# Conflicts:
#	src/main/java/ut/veikotiit/checkers/minimax/MtdF.java
#	src/main/java/ut/veikotiit/checkers/minimax/Negamax.java
#	src/main/java/ut/veikotiit/checkers/scorer/BitBoardScorer.java
#	src/main/java/ut/veikotiit/checkers/scorer/DefaultBitBoardScorer.java
#	src/main/java/ut/veikotiit/checkers/scorer/WeightedBitBoardScorer.java

+++ src/main/java/ut/veikotiit/checkers/scorer/BitBoardScorer.java
@@ -4,5 +4,5 @@ import ut.veikotiit.checkers.Color;
 import ut.veikotiit.checkers.bitboard.BitBoard;
 
 public interface BitBoardScorer {
-    int getScore(BitBoard bitBoard, Color color);
+    double getScore(BitBoard bitBoard, Color color);
 }
